Rename ui control config and csv layer
