+++ src/config/uicontrols.js
